4.21*: expansión de la lista de blogs, paso 9

+++ part4/bloglist/controllers/blogs.js
@@ -42,6 +42,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const user = await User.findById(decodedToken.id)
+  if (blog.user.toString() !== user._id.toString()) {
+    return response.status(401).json({
+      error: 'Unauthorized to delete this blog'
+    })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
